Select methods instead of filtering for them
Previously, the JUnit Platform test class processor always selected
classes and applied the filters as a `PostDiscoveryFilter`. That's
problematic in the presence of dynamic elements such as Spock `@Unroll`
methods which can be included or excluded that way. Now, we try to
derive method selectors from the filter expressions when running with
Vintage >= 5.7. This way, individual tests can be run from IDEs.